Check if the file exists prior removal

+++ src/PartKeepr/ImageBundle/Services/ImageService.php
@@ -45,7 +45,9 @@ class ImageService extends UploadedFileService
             /**
              * @var $file CachedImage
              */
-            unlink($file->getCacheFile());
+            if (file_exists($file->getCacheFile())) {
+                unlink($file->getCacheFile());
+            }
             $entityManager->remove($file);
         }
     }
